+++ VL_foxBot.lua
@@ -3227,7 +3227,7 @@ end, "game")
 ]]
 local function BotHelp(player)
 	print(
-		"\x87 foxBot! v1.4: 2021-xx-xx",
+		"\x87 foxBot! v1.4: 2021-08-20",
 		"\x81  Based on ExAI v2.0: 2019-12-31",
 		"",
 		"\x87 SP / MP Server Admin:",